# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The ChefConnect app is built using a modern Angular + back end API architecture. All API endpoints are detailed here. Many endpoints can be called as both an authenticated or unauthenticated user, though the data returned may differ. For example, when pulling back recipe details, the notes and ratings associated with that recipe will not be included unless authenticated.",
    "title": "ChefConnect API Docs",
    "contact": {
      "name": "atticuss",
      "email": "jonn.callahan@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost:8080",
  "paths": {
...
    "/ingredients": {
      "get": {
        "description": "Fetch all ingredients",
        "tags": [
          "ingredients"
        ],
        "operationId": "getAllIngredients",
        "responses": {
          "200": {
            "$ref": "#/responses/ManyIngredients"
          }
        }
      },
      "post": {
        "description": "Create a new ingredient",
        "tags": [
          "ingredients"
        ],
        "operationId": "createIngredient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "uid",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Amount",
            "name": "amount",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ingredient"
          }
        }
      }
    },
    "/ingredients/{id}": {
      "get": {
        "description": "Fetches a single ingredient by ID",
        "tags": [
          "ingredients"
        ],
        "operationId": "getIngredient",
        "responses": {
          "200": {
            "$ref": "#/responses/Ingredient"
          }
        }
      },
      "put": {
        "description": "Update an ingredient",
        "tags": [
          "ingredients"
        ],
        "operationId": "updateIngredient",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "uid",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Amount",
            "name": "amount",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/Ingredient"
          }
        }
      },
      "delete": {
        "description": "Delete an ingredient",
        "tags": [
          "ingredients"
        ],
        "operationId": "deleteIngredient",
        "responses": {}
      }
    },
...
    "/users": {
      "get": {
        "description": "Fetch all users",
        "tags": [
          "users"
        ],
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "$ref": "#/responses/ManyUsers"
          }
        }
      },
      "post": {
        "description": "Create a new user",
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "uid",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedRecipe"
            },
            "x-go-name": "Favorites",
            "name": "favorites",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedNote"
            },
            "x-go-name": "Notes",
            "name": "notes",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedRecipe"
            },
            "x-go-name": "Ratings",
            "name": "ratings",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Fetch all users",
        "tags": [
          "users"
        ],
        "operationId": "getUser",
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          }
        }
      },
      "put": {
        "description": "Update a user",
        "tags": [
          "users"
        ],
        "operationId": "updateUser",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "name": "uid",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedRecipe"
            },
            "x-go-name": "Favorites",
            "name": "favorites",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedNote"
            },
            "x-go-name": "Notes",
            "name": "notes",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/definitions/NestedRecipe"
            },
            "x-go-name": "Ratings",
            "name": "ratings",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/User"
          }
        }
      },
      "delete": {
        "description": "Delete a user",
        "tags": [
          "users"
        ],
        "operationId": "deleteUser",
        "responses": {}
      }
    }
  },
  "definitions": {
...
    "IngredientResponse": {
      "description": "IngredientResponse is a struct that represents a single ingredient. It is used exclusively\nfor marshalling responses back to API clients.",
      "type": "object",
      "properties": {
        "amount": {
          "type": "string",
          "x-go-name": "Amount"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NestedCategory"
          },
          "x-go-name": "Categories"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uid": {
          "type": "string",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
...
    "ManyIngredientsResponse": {
      "description": "ManyIngredientsResponse is a struct that represents multiple ingredients. It is used\nexclusively for marshalling responsesback to API clients.",
      "type": "object",
      "properties": {
        "ingredients": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserResponse"
          },
          "x-go-name": "Ingredients"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
...
    "ManyUsersResponse": {
      "description": "ManyUsersResponse is a struct that represents multiple users. It is used\nexclusively for marshalling responsesback to API clients.",
      "type": "object",
      "properties": {
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserResponse"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
    "NestedCategory": {
      "description": "NestedCategory is a stripped down struct used when a Category is nested\nwithin a parent struct in an API response",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uid": {
          "type": "string",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
...
    "NestedNote": {
      "description": "NestedNote is a stripped down struct used when a Note is nested\nwithin a parent struct in an API response",
      "type": "object",
      "properties": {
        "dgraph.type": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "DType"
        },
        "text": {
          "type": "string",
          "x-go-name": "Text"
        },
        "uid": {
          "type": "string",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
    "NestedRecipe": {
      "description": "NestedRecipe is a stripped down struct used when a Recipe is nested\nwithin a parent struct in an API response",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "uid": {
          "type": "string",
          "x-go-name": "ID"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    },
...
    "UserResponse": {
      "description": "UserResponse is a struct that represents a single user. It is used exclusively\nfor marshalling responses back to API clients.",
      "type": "object",
      "properties": {
        "favorites": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NestedRecipe"
          },
          "x-go-name": "Favorites"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "notes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NestedNote"
          },
          "x-go-name": "Notes"
        },
        "ratings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/NestedRecipe"
          },
          "x-go-name": "Ratings"
        },
        "uid": {
          "type": "string",
          "x-go-name": "ID"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "github.com/atticuss/chefconnect/models"
    }
  },
  "responses": {
...
    "Ingredient": {
      "schema": {
        "$ref": "#/definitions/IngredientResponse"
      }
    },
...
    "ManyIngredients": {
      "schema": {
        "$ref": "#/definitions/ManyIngredientsResponse"
      }
    },
...
    "ManyUsers": {
      "schema": {
        "$ref": "#/definitions/ManyUsersResponse"
      }
    },
...
    "User": {
      "schema": {
        "$ref": "#/definitions/UserResponse"
      }
    }
  }
}